refactor: moved datadog integration into integrations
