+++ app/integrations/datadog/correlation/__init__.py
@@ -1,0 +1,19 @@
+from __future__ import annotations
+
+from app.integrations.datadog.correlation.adapter import DatadogCorrelationAdapter
+from app.integrations.datadog.correlation.factory import (
+    build_datadog_provider,
+    datadog_avg_query,
+)
+from app.integrations.datadog.correlation.provider import (
+    DatadogCorrelationQueries,
+    DatadogUpstreamEvidenceProvider,
+)
+
+__all__ = [
+    "DatadogCorrelationAdapter",
+    "DatadogCorrelationQueries",
+    "DatadogUpstreamEvidenceProvider",
+    "build_datadog_provider",
+    "datadog_avg_query",
+]
