+++ app/integrations/datadog/correlation/provider.py
@@ -2,13 +2,11 @@ from __future__ import annotations
 
 from dataclasses import dataclass
 
-from app.agent.stages.publish_findings.upstream_correlation.datadog_adapter import (
-    DatadogCorrelationAdapter,
-)
 from app.agent.stages.publish_findings.upstream_correlation.upstream import (
     TopologyHint,
     UpstreamEvidenceBundle,
 )
+from app.integrations.datadog.correlation.adapter import DatadogCorrelationAdapter
 
 
 @dataclass(frozen=True)
