+++ tests/test_datadog_provider.py
@@ -1,9 +1,9 @@
 from __future__ import annotations
 
-from app.agent.stages.publish_findings.upstream_correlation.datadog_adapter import (
+from app.integrations.datadog.correlation.adapter import (
     DatadogCorrelationAdapter,
 )
-from app.agent.stages.publish_findings.upstream_correlation.datadog_provider import (
+from app.integrations.datadog.correlation.provider import (
     DatadogCorrelationQueries,
     DatadogUpstreamEvidenceProvider,
 )
